+++ py-polars/src/lib.rs
@@ -11,7 +11,6 @@ mod build {
 }
 
 mod allocator;
-mod arrow_interop;
 #[cfg(feature = "csv")]
 mod batched_csv;
 mod conversion;
@@ -22,6 +21,7 @@ mod expr;
 mod file;
 mod functions;
 mod gil_once_cell;
+mod interop;
 mod lazyframe;
 mod lazygroupby;
 mod map;
