+++ py-polars/src/series/construction.rs
@@ -10,10 +10,10 @@ use polars_core::utils::CustomIterTools;
 use pyo3::exceptions::PyTypeError;
 use pyo3::prelude::*;
 
-use crate::arrow_interop::to_rust::array_to_rust;
 use crate::conversion::any_value::py_object_to_any_value;
 use crate::conversion::Wrap;
 use crate::error::PyPolarsErr;
+use crate::interop::arrow::to_rust::array_to_rust;
 use crate::prelude::ObjectValue;
 use crate::PySeries;
